Set up Handling for Positive adjusts and cleaning up the logic for negative adjusts.

+++ lib/ADWatch_GPS/src/GPSTools.h
@@ -22,7 +22,7 @@ class GPSTools{
 	public:
 		//Attributes
 		Adafruit_GPS* gps;
-		int8_t prev_adjust;
+		int prev_adjust;
 		//Methods
 		GPSTools(Adafruit_GPS*);
 		time_t grabTime(Adafruit_GPS);
